// src/type/types.ts
export type Type =
  | 'unknown'
  | 'undefined'
  | 'null'
  | 'number'
  | 'string'
  | 'boolean'
  | 'object'
  | 'symbol'
  | 'Array'
  | 'Date'
  | 'RegExp'
  | 'Set'
  | 'Map'
  | 'WeakSet'
  | 'WeakMap'
  | 'Promise'
  | 'Error'
  | 'Int8Array'
  | 'Uint8Array'
  | 'Uint8ClampedArray'
  | 'Int16Array'
  | 'Uint16Array'
  | 'Int32Array'
  | 'Uint32Array'
  | 'Float32Array'
  | 'Float64Array'
  | 'BigInt64Array'
  | 'BigUint64Array'
  | 'ArrayBuffer'
  | string;